requirements.txt
streamlit
paho-mqtt
pandas
numpy
joblib
plotly
google-generativeai
python-dotenv
scikit-learn